pyproject.toml
[tool.poetry]
name = "automatica-test-task"
version = "0.1.0"
description = "This project is a test task completed as part of a job interview process for Automatica."
authors = ["smbrine <smbrine@yandex.ru>"]
license = "GPL-3.0-or-later"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
django = "4.2"
djangorestframework = "^3.15.2"
pydantic = "^2.8.2"
psycopg2-binary = "^2.9.9"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
pylint = "^3.2.5"
flake8 = "^7.1.0"
ruff = "^0.5.2"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
